requirements.txt
alabaster==0.7.12
amqp==1.4.9
anyjson==0.3.3
apns2==0.7.1
appdirs==1.4.4

asgiref==3.2.7
asn1crypto==0.24.0
auditlog3==1.0.1
Babel==2.9.0
backcall==0.1.0
beautifulsoup4==4.9.1
billiard==3.3.0.23
braintree==4.1.0

bs4==0.0.1
celery==3.1.26.post2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.3.7

coverage==5.5
cryptography==2.9.2

decorator==4.4.2

defusedxml==0.6.0
distlib==0.3.1

dj-database-url==0.3.0
Django==3.0.6
django-allauth==0.44.0
django-celery==3.3.1
django-ckeditor==5.9.0
django-clever-selects==0.8.2
django-countries-plus==1.2.1
django-currencies==0.10.1
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==3.1.1
django-filer==1.3.0
django-grappelli==2.14.3
django-jet==1.0.7
django-js-asset==1.2.2
django-map-widgets==0.1.9
django-mptt==0.8.7
django-polymorphic==1.0.2
django-push-notifications==2.0.0
django-rest-auth==0.9.3
django-utils-six==2.0
django-uuid-upload-path==1.0.0
django-versatileimagefield==1.6
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
docutils==0.16
easy-thumbnails==2.3
filelock==3.0.12
flake8==3.9.0

geographiclib==1.50
geopy==1.22.0
gunicorn==20.0.4
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.9.2
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
importlib-resources==5.1.2
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.3
jsonfield==2.1.1
keyring==10.6.0
keyrings.alt==3.0
kombu==3.0.37

launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3

macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2

mccabe==0.6.1
netifaces==0.10.4
numpy==1.18.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.44
packaging==20.9
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==4.3.0
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.0.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.10.0
py-vapid==1.7.0
pycairo==1.16.2
pycodestyle==2.7.0
pycountry==19.8.18
pycparser==2.20
pycrypto==2.6.1
# pycups==1.9.73
pyfcm==1.4.7
pyflakes==2.3.0
Pygments==2.1.3
# pygobject==3.27.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0

python-dateutil==2.8.1
python-debian==0.1.32
python-quickbooks==0.7.4
python3-openid==3.1.0
pytz==2020.1
pywebpush==1.11.0
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
rauth==0.7.3
redis==2.10.6
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
reverse-geocoder==1.5.1
scipy==1.4.1
SecretStorage==2.3.1
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.1.0
soupsieve==2.0.1
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4


toml==0.10.2
tox==3.23.0
traitlets==4.3.3
typing==3.7.4.1


Unidecode==0.4.21
urllib3==1.25.9


virtualenv==20.4.3
wadllib==1.3.2
wcwidth==0.1.9
whitenoise==5.0.1
boto
boto3
django-storages==1.11.1
